Trait recur_fn::RecurFn

pub trait RecurFn<Arg, Output> {
    // Required method
    fn body(&self, recur: impl Fn(Arg) -> Output, arg: Arg) -> Output;

    // Provided method
    fn call(&self, arg: Arg) -> Output { ... }
}

The recursive function trait.

Instead of recurring directly, this trait allows user to customize the recursion. In this way, we can extract and extend the body of a recursive function.

This trait supports only one argument. If you need multiple arguments, use tuple.

Required Methods

fn body(&self, recur: impl Fn(Arg) -> Output, arg: Arg) -> Output

The body of the recursive function.

Marking this method by #[inline] will make the call method as fast as recurring directly.

Provided Methods

fn call(&self, arg: Arg) -> Output

Calls the recursive function.

Object Safety

This trait is not object safe.